srcs/backend/requirements.txt
asgiref==3.7.2
certifi @ file:///private/tmp/python-certifi-20231008-5561-13s00fm/certifi-2023.7.22
cffi @ file:///private/tmp/cffi-20231014-5567-1hyue8f/cffi-1.16.0
charset-normalizer==3.3.2
cryptography @ file:///private/tmp/python-cryptography-20231024-6375-1k9sys9/cryptography-41.0.5
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
itsdangerous==2.1.2
lxml @ file:///private/tmp/python-lxml-20231020-5438-v8d9ch/lxml-4.9.3
MarkupSafe @ file:///private/tmp/python-markupsafe-20231006-5545-ondi82/MarkupSafe-2.1.3
packaging @ file:///private/tmp/python-packaging-20231018-5368-zv724j/packaging-23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser @ file:///private/tmp/pycparser-20231013-5787-1anpge2/pycparser-2.21
pycryptosat @ file:///private/tmp/cryptominisat-20230811-6541-1bndh55/cryptominisat-5.11.12
PyJWT==1.7.1
pytesseract==0.3.10
pytz @ file:///private/tmp/python-pytz-20231024-5403-w4fpxw/pytz-2023.3.post1
PyYAML @ file:///private/tmp/pyyaml-20231018-5610-1rdh755/PyYAML-6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
